Split workspace asset hooks
